tokyo-crafter-infra-11 setup.sh が冪等になるようにリファクタリング

+++ proxy/modules/aws/lightsail/setup.sh
@@ -1,41 +1,53 @@
 #!/bin/bash
 
 ###### install docker
-echo "install docker"
-sudo rm -f /etc/yum.repos.d/docker-ce.repo
-sudo yum update -y
+if [ -x "$(command -v docker)" ]; then
+  echo "docker is installed"
+  echo "skip installing docker"
 
-sudo yum install yum-utils amazon-linux-extras -y
-sudo amazon-linux-extras install docker -y
+else
+  echo "docker is not installed"
 
-sudo groupadd docker
-sudo usermod -aG docker $USER
+  echo "install docker"
+  sudo rm -f /etc/yum.repos.d/docker-ce.repo
+  sudo yum update -y
 
-sudo systemctl enable docker
-sudo systemctl enable containerd
-sudo systemctl start containerd
-sudo systemctl status containerd
-sudo systemctl start docker
-sudo systemctl status docker
+  sudo yum install yum-utils amazon-linux-extras -y
+  sudo amazon-linux-extras install docker -y
 
-###### install docker-compose
-echo "install docker-compose"
-sudo yum install curl
-sudo curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-sudo chmod +x /usr/local/bin/docker-compose
-sudo ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose
-docker-compose --version
+  sudo groupadd docker
+  sudo usermod -aG docker $USER
+
+  sudo systemctl enable docker
+  sudo systemctl enable containerd
+  sudo systemctl start containerd
+  sudo systemctl status containerd
+  sudo systemctl start docker
+  sudo systemctl status docker
+
+  ###### install docker-compose
+  echo "install docker-compose"
+  sudo yum install curl
+  sudo curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+  sudo chmod +x /usr/local/bin/docker-compose
+  sudo ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose
+  docker-compose --version
+
+fi
 
 ###### install vpn client
-echo "setup vpn"
-sudo amazon-linux-extras install epel -y
-sudo yum install strongswan xl2tpd -y
+if [[ $(ping "$VPN_TARGET_IP_ID" -c 1) == *"1 received"* ]]; then
+  echo "vpn already connected"
+else
+  echo "setup vpn"
+  sudo amazon-linux-extras install epel -y
+  sudo yum install strongswan xl2tpd -y
 
-sudo tee -a /etc/strongswan/ipsec.conf <<EOF >/dev/null
+  sudo tee -a /etc/strongswan/ipsec.conf <<EOF >/dev/null
 include /etc/strongswan/ipsec.d/*.conf
 EOF
 
-sudo tee -a /etc/strongswan/ipsec.d/yamaha.conf <<EOF >/dev/null
+  sudo tee /etc/strongswan/ipsec.d/yamaha.conf <<EOF >/dev/null
 
 conn %default
     keyingtries=0
@@ -80,11 +92,11 @@ conn yamaha
 
 EOF
 
-sudo tee -a /etc/strongswan/ipsec.secrets <<EOF >/dev/null
+  sudo tee /etc/strongswan/ipsec.secrets <<EOF >/dev/null
 : PSK "$VPN_PSK"
 EOF
 
-sudo tee /etc/xl2tpd/xl2tpd.conf <<EOF >/dev/null
+  sudo tee /etc/xl2tpd/xl2tpd.conf <<EOF >/dev/null
 [lac yamaha]
 lns = $VPN_TARGET_IP
 require chap = yes
@@ -99,7 +111,7 @@ max redials = 10
 
 EOF
 
-sudo tee /etc/ppp/options.xl2tpd.yamaha <<EOF >/dev/null
+  sudo tee /etc/ppp/options.xl2tpd.yamaha <<EOF >/dev/null
 # L2TPトンネルのユーザ名
 name $VPN_USER
 
@@ -121,30 +133,30 @@ logfile /var/log/ppp/yamaha.log
 
 EOF
 
-sudo tee /etc/ppp/chap-secrets <<EOF >/dev/null
+  sudo tee /etc/ppp/chap-secrets <<EOF >/dev/null
 $VPN_USER    *    $VPN_PASSWORD    *
 EOF
 
-sudo tee /etc/ppp/pap-secrets <<EOF >/dev/null
+  sudo tee /etc/ppp/pap-secrets <<EOF >/dev/null
 $VPN_USER    *    $VPN_PASSWORD    *
 EOF
 
-sudo systemctl enable strongswan
-sudo systemctl start strongswan
+  sudo systemctl enable strongswan
+  sudo systemctl start strongswan
 
-sleep 10
+  sleep 10
 
-sudo systemctl enable xl2tpd
-sudo systemctl start xl2tpd
+  sudo systemctl enable xl2tpd
+  sudo systemctl start xl2tpd
 
-sleep 20
+  sleep 20
 
-sudo xl2tpd-control connect yamaha
+  sudo xl2tpd-control connect yamaha
 
-sleep 30
+  sleep 30
 
-sudo mkdir /etc/cron.every_minute
-sudo tee /etc/cron.d/0every_minute <<EOF >/dev/null
+  sudo mkdir /etc/cron.every_minute
+  sudo tee /etc/cron.d/0every_minute <<EOF >/dev/null
 SHELL=/bin/bash
 PATH=/sbin:/bin:/usr/sbin:/usr/bin
 MAILTO=root
@@ -152,31 +164,37 @@ HOME=/
 * * * * * root run-parts /etc/cron.every_minute
 EOF
 
-sudo mkdir /var/log/xl2tpd
-sudo touch /var/log/xl2tpd/keepalive.log
-sudo tee /etc/cron.every_minute/keepalive_vpn.sh <<EOF >/dev/null
+  sudo mkdir /var/log/xl2tpd
+  sudo touch /var/log/xl2tpd/keepalive.log
+  sudo tee /etc/cron.every_minute/keepalive_vpn.sh <<EOF >/dev/null
 
 #!/bin/sh
 echo $(date) > /var/log/xl2tpd/keepalive.log
 ping $VPN_TARGET_IP_ID -c 1 >> /var/log/xl2tpd/keepalive.log
 
 EOF
-sudo chmod a+x /etc/cron.every_minute/keepalive_vpn.sh
+  sudo chmod a+x /etc/cron.every_minute/keepalive_vpn.sh
 
-vpn_connected=false
-for i in {1..100}; do
+  vpn_connected=false
+  for i in {1..100}; do
 
-  if cat /var/log/xl2tpd/keepalive.log | grep "1 received"; then
-    vpn_connected=true
-  fi
+    if cat /var/log/xl2tpd/keepalive.log | grep "1 received"; then
+      vpn_connected=true
+    fi
 
-  if [ $vpn_connected == true ]; then
-    break
-  else
-    echo "waiting...($i)"
-  fi
-  sleep 10
-done
-echo "vpn connected"
+    if [ $vpn_connected == true ]; then
+      break
+    else
+      echo "waiting...($i)"
+      if [ $i == 100 ]; then
+        echo "failed to connect"
+        exit 1
+      fi
+    fi
+    sleep 10
+  done
+  echo "vpn connected"
+
+fi
 
 exit 0
